refactor pixelator_maxiv to pixelator_common
