+++ cls/applications/pyStxm/bl_configs/maxiv_pixelator/scan_plugins/line_scan/line_scan.py
@@ -13,12 +13,10 @@ from cls.utils.roi_dict_defs import *
 from cls.applications.pyStxm.bl_configs.base_scan_plugins.line_scan.line_scan import (
     BaseLineScanParam,
 )
-from cls.applications.pyStxm.bl_configs.maxiv_pixelator.plugin_utils import (connect_scan_req_detail_flds_to_validator,
-    init_scan_req_member_vars)
 from cls.utils.roi_utils import wdg_to_sp, get_sp_db_from_wdg_com
 
-from cls.applications.pyStxm.bl_configs.maxiv_pixelator.plugin_utils import (connect_scan_req_detail_flds_to_validator,
-    init_scan_req_member_vars, set_scan_rec_default, scan_rec_enable_widget)
+from cls.applications.pyStxm.bl_configs.pixelator_common.plugin_utils import (connect_scan_req_detail_flds_to_validator,
+                                                                              init_scan_req_member_vars, set_scan_rec_default, scan_rec_enable_widget)
 
 _logger = get_module_logger(__name__)
